fix, feat, chore: pretty much everything left
This commit includes a lot of stuff that was done during the past two weeks. That includes some redesinging, fixing major bugs in structure/logic, reformatting the code that was previously there and adding a few windows for easier use. I understand that this may not be the best solution to dump this in a single commit, but this makes more sense than to leave out unfinished pieces together. The only things left are some more debugging and synchronizing every mounting in the MainWindow class

Signed-off-by: Thomas Gauthier <thomgau17@gmail.com>