fix: threading and query
Fixed the querying, which had a few bugs in it. Also, learned that the multiprocessing module copies the environment of the virtual machine, which implies that a switch was required from pocesses to threads. Also, PyQt does not allow another thread to modify the UI, so a fix (hack) has been created to circumvent this issue. This was partaken during the last two weeks, hence the lack of commits

Signed-off-by: Thomas Gauthier <thomgau17@gmail.com>